add script for creating table users-products and fix script

+++ Database/DbSctipts/BadaboomDB/2_create_table_users.sql
@@ -1,5 +1,5 @@
 ﻿create table [Users](
-	[UserId] BIGINT NOT NULL IDENTITY(1,1) PRIMARY KEY,
+	[UserId] INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
 
 	[Address] binary(20) NOT NULL UNIQUE,
